Refactor SubscriptionPriceCard
