+++ dashboard/src/components/subscriptions/SubscriptionStripePriceCard.tsx
@@ -1,16 +1,24 @@
+import { faStripe } from "@fortawesome/free-brands-svg-icons";
+import {
+  faCircleCheck,
+  faCircleQuestion,
+  faPeopleGroup,
+  faPerson,
+} from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { getSubscriptionPriceBanner } from "helpers/subscriptions";
 import { getCurrency } from "helpers/utils";
-import { Card } from "react-bootstrap";
+import { Button, Card, ListGroup } from "react-bootstrap";
 import { SubscriptionPrice } from "types/subscription";
 
 interface SubscriptionPriceCardProps {
   price: SubscriptionPrice;
-  children?: JSX.Element | string | number | false;
+  onSelect?: (priceId: string) => void;
 }
 
 const SubscriptionStripePriceCard = ({
   price,
-  children,
+  onSelect,
 }: SubscriptionPriceCardProps) => {
   return (
     <Card>
@@ -28,8 +36,41 @@ const SubscriptionStripePriceCard = ({
           </div>
         </div>
       </Card.Body>
-      {children && <hr className="mx-2" />}
-      {children}
+      <hr className="mx-2" />
+      <Card.Body className="pt-0">
+        <ListGroup>
+          <ListGroup.Item className="d-flex align-items-center">
+            <FontAwesomeIcon icon={faCircleCheck} className="s-1" />
+            <span className="ps-2">No ads</span>
+          </ListGroup.Item>
+          <ListGroup.Item className="d-flex align-items-center">
+            <FontAwesomeIcon icon={faPerson} className="s-1" />
+            <span className="ps-2">10 Player slots</span>
+          </ListGroup.Item>
+          <ListGroup.Item className="d-flex align-items-center">
+            <FontAwesomeIcon icon={faPeopleGroup} className="s-1" />
+            <span className="ps-2">1 Guild slot</span>
+          </ListGroup.Item>
+          <ListGroup.Item className="d-flex align-items-center">
+            <FontAwesomeIcon icon={faCircleQuestion} className="s-1" />
+            <span className="ps-2">Premium support</span>
+          </ListGroup.Item>
+        </ListGroup>
+      </Card.Body>
+
+      <Card.Footer>
+        <div className="d-flex justify-content-stretch">
+          <Button
+            variant="primary"
+            style={{ width: "100%" }}
+            className="d-flex align-items-center"
+            onClick={() => onSelect && onSelect(price.id)}
+          >
+            <FontAwesomeIcon icon={faStripe} className="s-2" />
+            <div>Checkout</div>
+          </Button>
+        </div>
+      </Card.Footer>
     </Card>
   );
 };
